+++ src/lvl_script_commands_old.h
@@ -23,7 +23,6 @@
 extern "C" {
 #endif
 
-#include "lvl_script_lib.h"
 #include "lvl_script_commands.h"
 
 void script_add_command(const struct CommandDesc *cmd_desc, const struct ScriptLine *scline, long file_version);
