+++ src/lvl_script_lib.c
@@ -198,7 +198,7 @@ long get_players_range_single_f(long plr_range_id, const char *func_name, long l
     return -2;
 }
 
-void get_player_number_from_value(const char* txt, char* id, char* type)
+void get_chat_icon_from_value(const char* txt, char* id, char* type)
 {
     char idx;
     if (strcasecmp(txt, "None") == 0)
