+++ src/lvl_script_lib.h
@@ -295,7 +295,27 @@ enum ScriptVariables {
 
 
 
-
+extern const struct NamedCommand player_desc[];
+extern const struct NamedCommand controls_variable_desc[];
+extern const struct NamedCommand timer_desc[];
+extern const struct NamedCommand flag_desc[];
+extern const struct NamedCommand hand_rule_desc[];
+extern const struct NamedCommand rule_slot_desc[];
+extern const struct NamedCommand rule_action_desc[];
+extern const struct NamedCommand hero_objective_desc[];
+extern const struct NamedCommand msgtype_desc[];
+extern const struct NamedCommand tendency_desc[];
+extern const struct NamedCommand creature_select_criteria_desc[];
+extern const struct NamedCommand trap_config_desc[];
+extern const struct NamedCommand gui_button_group_desc[];
+extern const struct NamedCommand campaign_flag_desc[];
+extern const struct NamedCommand script_operator_desc[];
+extern const struct NamedCommand variable_desc[];
+extern const struct NamedCommand dk1_variable_desc[];
+extern const struct NamedCommand fill_desc[];
+extern const struct NamedCommand set_door_desc[];
+extern const struct NamedCommand texture_pack_desc[];
+extern const struct NamedCommand locked_desc[];
 
 // 1/4 turn minimal
 #define FX_LINE_TIME_PARTS 4
@@ -315,7 +335,7 @@ long parse_criteria(const char *criteria);
 #define get_players_range_single(plr_range_id) get_players_range_single_f(plr_range_id, __func__, text_line_number)
 long get_players_range_single_f(long plr_range_id, const char *func_name, long ln_num);
 TbBool parse_get_varib(const char *varib_name, long *varib_id, long *varib_type);
-void get_player_number_from_value(const char* txt, char* id, char* type);
+void get_chat_icon_from_value(const char* txt, char* id, char* type);
 #define get_player_id(plrname, plr_range_id) get_player_id_f(plrname, plr_range_id, __func__, text_line_number)
 TbBool get_player_id_f(const char *plrname, long *plr_range_id, const char *func_name, long ln_num);
 TbResult script_use_power_on_creature(struct Thing* thing, short pwkind, KeepPwrLevel power_level, PlayerNumber caster, TbBool is_free);
