+++ src/thing_stats.h
@@ -71,8 +71,9 @@ typedef short CreatureLiveStatId;
 extern const char *blood_types[15];
 /******************************************************************************/
 const char *thing_model_name(const struct Thing *thing);
-const char *thing_class_and_model_name(int class_id, int model);
-const char *thing_class_code_name(int class_id);
+const char *thing_model_only_name(ThingClass class_id, ThingModel model);
+const char *thing_class_and_model_name(ThingClass class_id, ThingModel model);
+const char *thing_class_code_name(ThingClass class_id);
 const char *creatrtng_actstate_name(const struct Thing *thing);
 const char *creatrtng_realstate_name(const struct Thing *thing);
 TbBool things_stats_debug_dump(void);
